use separate test packages in registry and manager
Since these packages want to use testcommon for their tests, but
testcommon also wants to use the registry and manager packages, we have
an import cycle.

The simplest fix is to make the manager and registry tests sit under
separate _test packages, so that manager_test is entirely separate from
the manager package. This way, we no longer have an import cycle.

The only downside is that, since manager_test is a separate package, it
can't access unexported functions or variables in the manager package.
But that seems like an OK tradeoff, given that these look like
integration tests.

Analternative would be to put these tests in the "test" package along
with other integration tests. It doesn't seem like a clearly better or
worse option, and it's a larger refactor, so we're not doing it here.