fix ORDER BY param in ListMembers
ORDER BY does not accept dynamic params so fmt.Sprintf should
be used. But there is a compromise with this approach
and is that it opens the door to SQL injections.
In order to prevent undesired behaviours, the order by
accepted strings are now consts with iotas and only
the MemberInfo keys are accepted.