Add news feed push notifications:
The news feed push notifications informs a group of users subscribed to a topic that an entity created/modified its news feed.

Given that the news feed data is stored on IPFS and the IPFS URI is stored on an EVM compatible chain, an IPFS and an Ethereum endpoint are required (internal or external). It is also required to have access to the manager database in order to get an updated entity list.

In order to track the entities news feed (part of the entity metadata) this commit introduces the IPFSFileTracker service. This service runs forever and is in charge of:

- Refreshing entity list using the local database.
- Tracking each entity metadata looking for changes on the news feed.

The first step is to refresh the list of entities on each iteration and then retrieve the entity metadata.
For each entity on the list (for each refresh) a new go routine with a timeout is created and added to a wait group, so the iteration will finish if one of the following scenarios is reached:

- All entities refreshed.
- Timeout reached. This is needed because of the IPFS data availability times.

The go routines fetch the new entity metadata and compare the new data with the old data stored by the tracker. If the data is different the tracker updates its data stored and created a push notification to be handled by the notifier service.