add ethereum faucet to the manager:
Currently when an entity signs up its admins
need to request the tokens required for interacting
with the undelying EVM blockchain via email.
This process does not scale and adds friction to the user.

For solving that, this commit enables the manager to
do so automatically when an entity creates an account.
Also allows an entity which already spend its tokens
to ask for more tokens in order to continue its operations.

The first batch of tokens will be sended automatically
on the entity creation and for requesting more tokens
the entity just need to call the /requestTokens manager endpoint.

- added ethclient to manager
- added Signer() method to get the ethereum.SigningKey from the router
- added default token amount to send, gas limit and the threshold
  required to request more tokens as const
- added /requestTokens endpoint
- added flag for setting the web3 endpoint
- add sendTokens method to the manager
- modify tests accordingly