Feature/signers pool (#40)
faucet: add signers pool
Add a pool of signers for the faucet. If Different 
signup or request gas requests are received
in a short period of time, the signer can be blocked
on a pending blockchain transaction.
Adding the pool of signers allows to use a free
signer that does not have pending tx's.